Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
reportlab
pymysql
cryptography
